Create a new sycamore-core crate
