+++ packages/sycamore-core/src/lib.rs
@@ -1,0 +1,6 @@
+//! Core functionality for the Sycamore web framework.
+
+pub mod generic_node;
+pub mod noderef;
+pub mod render;
+pub mod view;
