+++ packages/sycamore-core/src/render.rs
@@ -3,10 +3,9 @@
 use std::rc::Rc;
 
 use ahash::AHashMap;
-use wasm_bindgen::UnwrapThrowExt;
+use sycamore_reactive::*;
 
 use crate::generic_node::GenericNode;
-use crate::reactive::*;
 use crate::view::{View, ViewType};
 
 /// Insert a [`GenericNode`] under `parent` at the specified `marker`. If `initial` is `Some(_)`,
@@ -291,7 +290,7 @@ pub fn reconcile_fragments<G: GenericNode>(parent: &G, a: &mut [G], b: &[G]) {
         } else if b_end == b_start {
             // Remove.
             while a_start < a_end {
-                if map.is_none() || !map.as_ref().unwrap_throw().contains_key(&a[a_start]) {
+                if map.as_ref().map_or(false, |m| m.contains_key(&a[a_start])) {
                     parent.remove_child(&a[a_start]);
                 }
                 a_start += 1;
@@ -318,12 +317,14 @@ pub fn reconcile_fragments<G: GenericNode>(parent: &G, a: &mut [G], b: &[G]) {
         } else {
             // Fallback to map.
             if map.is_none() {
-                map = Some(AHashMap::with_capacity(b_end - b_start));
-                for (i, item) in b.iter().enumerate().take(b_end).skip(b_start) {
-                    map.as_mut().unwrap_throw().insert(item.clone(), i);
-                }
+                let tmp = b[b_start..b_end]
+                    .iter()
+                    .enumerate()
+                    .map(|(i, g)| (g.clone(), i))
+                    .collect();
+                map = Some(tmp);
             }
-            let map = map.as_ref().unwrap_throw();
+            let map = map.as_ref().unwrap();
 
             let index = map.get(&a[a_start]);
             if let Some(index) = index {
@@ -335,7 +336,7 @@ pub fn reconcile_fragments<G: GenericNode>(parent: &G, a: &mut [G], b: &[G]) {
                     while i + 1 < a_end && i + 1 < b_end {
                         i += 1;
                         t = map.get(&a[i]);
-                        if t.is_none() || *t.unwrap_throw() != *index + sequence {
+                        if t.is_none() || *t.unwrap() != *index + sequence {
                             break;
                         }
                         sequence += 1;
