+++ packages/sycamore/src/futures.rs
@@ -14,10 +14,10 @@
 
 use std::future::Future;
 
-// Re-export `wasm_bindgen_futures` crate.
-pub use wasm_bindgen_futures::*;
 // Re-export `sycamore-futures` crate.
 pub use sycamore_futures::*;
+// Re-export `wasm_bindgen_futures` crate.
+pub use wasm_bindgen_futures::*;
 
 use crate::prelude::*;
