+++ packages/sycamore/src/generic_node/hydrate_dom.rs
@@ -7,15 +7,14 @@ use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
 use web_sys::Node;
 
+use super::dom_node::NodeId;
+use super::SycamoreElement;
 use crate::generic_node::{DomNode, GenericNode, Html};
 use crate::reactive::*;
 use crate::utils::hydrate::web::get_next_element;
 use crate::utils::hydrate::{hydration_completed, with_hydration_context};
 use crate::utils::render::insert;
 use crate::view::View;
-
-use super::dom_node::NodeId;
-use super::SycamoreElement;
 
 /// Rendering backend for the DOM with hydration support.
 ///
@@ -84,6 +83,8 @@ impl fmt::Debug for HydrateNode {
 
 impl GenericNode for HydrateNode {
     type EventType = web_sys::Event;
+    type PropertyType = JsValue;
+
     const USE_HYDRATION_CONTEXT: bool = true;
     const CLIENT_SIDE_HYDRATION: bool = true;
