+++ packages/sycamore/src/motion.rs
@@ -4,7 +4,8 @@ use std::cell::{Cell, RefCell};
 use std::rc::Rc;
 
 use js_sys::Date;
-use wasm_bindgen::{prelude::*, JsCast};
+use wasm_bindgen::prelude::*;
+use wasm_bindgen::JsCast;
 
 use crate::reactive::*;
