+++ packages/sycamore/src/utils/mod.rs
@@ -9,16 +9,17 @@ use crate::prelude::View;
 
 #[cfg(feature = "hydrate")]
 pub mod hydrate;
-pub mod render;
+pub use sycamore_core::render;
 
 /// If `el` is a `HydrateNode`, use `get_next_marker` to get the initial node value.
 pub fn initial_node<G: GenericNode>(_el: &G) -> Option<View<G>> {
     #[cfg(feature = "hydrate")]
     {
-        use crate::generic_node::HydrateNode;
         use std::any::Any;
         use std::mem::ManuallyDrop;
         use std::ptr;
+
+        use crate::generic_node::HydrateNode;
         if let Some(el) = <dyn Any>::downcast_ref::<HydrateNode>(_el) {
             let initial = hydrate::web::get_next_marker(&el.inner_element());
             // Do not drop the HydrateNode because it will be cast into a GenericNode.
