+++ packages/sycamore/tests/web/reconcile.rs
@@ -5,10 +5,9 @@
 use sycamore::generic_node::DomNode;
 use sycamore::html;
 use sycamore::utils::render::{append_nodes, insert, reconcile_fragments};
+use wasm_bindgen_test::*;
 
 use super::*;
-
-use wasm_bindgen_test::*;
 
 #[wasm_bindgen_test]
 fn insert_create_nodes() {
