+++ rustfmt.toml
@@ -1,2 +1,5 @@
 comment_width = 100
 wrap_comments = true
+imports_granularity = "Module"
+reorder_imports = true
+group_imports = "StdExternalCrate"
